+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.5.0 (in progress)
 -------------------
 * **(breaking change)** Method `.tree()` renamed to `.inspectTree()`.
+* Added aliases to `fs.createReadStream` and `fs.createWriteStream`.
 * Added more telling error when path to copy() desn't exist.
 
 0.4.1 (2014-07-16)